Disposal fixes + Apple metal GC fix
