+++ src/Maui/DrawnUi/Platforms/MacCatalyst/ConfigureHandlers.Catalyst.cs
@@ -6,10 +6,10 @@ namespace DrawnUi.Draw
     {
         public static void ConfigureHandlers(IMauiHandlersCollection handlers)
         {
-            handlers.AddHandler(typeof(SkiaViewAccelerated), typeof(SKGLViewHandlerRetained));
             handlers.AddHandler(typeof(MauiEntry), typeof(MauiEntryHandler));
             handlers.AddHandler(typeof(MauiEditor), typeof(MauiEditorHandler));
             handlers.AddHandler(typeof(DrawnUiBasePage), typeof(DrawnUiBasePageHandler));
+            handlers.AddHandler(typeof(SkiaViewAccelerated), typeof(SKGLViewHandlerRetained));
         }
     }
 }
